+++ Game/Snapshots/PlayerSnapshot.cs
@@ -4,7 +4,7 @@ namespace block_racing_common.Game.Snapshots
 {
     public class PlayerSnapshot
     {
-        public int Id { get; }
+        public long Id { get; }
 
         public int CarX { get; }
 
@@ -25,7 +25,7 @@ namespace block_racing_common.Game.Snapshots
         public LaneSnapshot Lane { get; }
 
         
-        public PlayerSnapshot(int id, int carX, float distance,
+        public PlayerSnapshot(long id, int carX, float distance,
             float speed, bool isStunned, PlayMode mode, PieceType? currentPieceType,
             Rotation? currentPieceRotation, float shootCooldownRemaining,
             LaneSnapshot lane)
